Fix: Refactor LinkedIn scheduler to use a dedicated cron endpoint
The LinkedIn scheduler was failing due to an `Invalid or missing x-vercel-cron-secret header` error. This indicates an issue with the Vercel-injected cron secret.

This commit refactors the scheduler to use a new, dedicated cron job endpoint at `/api/cron/linkedin.js`. This endpoint is secured with a `CRON_SECRET` environment variable, which provides a more robust and explicit authentication mechanism for the cron job.

The original scheduler logic has been moved from `api/schedule.js` to the new endpoint. The `api/schedule.js` file now only handles user-facing scheduling actions.

The `vercel.json` file has been updated to point the cron job to the new endpoint.

Additionally, the corresponding test file `test/test_scheduler_success.test.js` has been updated to reflect these changes. The test now mocks `node-fetch` to avoid actual network requests and no longer depends on a live database, making it more reliable.
