+++ vercel.json
@@ -1,7 +1,7 @@
 {
   "crons": [
     {
-      "path": "/api/schedule",
+      "path": "/api/cron/linkedin",
       "schedule": "0 8-23 * * *"
     }
   ],
